Test endpoint for OrganisationProfileEndpoint
